// src/lib/openapi/squad/models/OpportunityRelationshipsPayload.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Squad API
 * API for managing Squad resources
 *
 * The version of the OpenAPI document: 0.14.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime.js';
/**
 * Request body for managing opportunity relationships
 * @export
 * @interface OpportunityRelationshipsPayload
 */
export interface OpportunityRelationshipsPayload {
    /**
     * Array of solution IDs to associate with the opportunity
     * @type {Array<string>}
     * @memberof OpportunityRelationshipsPayload
     */
    solutionIds?: Array<string>;
    /**
     * Array of outcome IDs to associate with the opportunity
     * @type {Array<string>}
     * @memberof OpportunityRelationshipsPayload
     */
    outcomeIds?: Array<string>;
    /**
     * Array of feedback IDs to associate with the opportunity
     * @type {Array<string>}
     * @memberof OpportunityRelationshipsPayload
     */
    feedbackIds?: Array<string>;
}

/**
 * Check if a given object implements the OpportunityRelationshipsPayload interface.
 */
export function instanceOfOpportunityRelationshipsPayload(value: object): value is OpportunityRelationshipsPayload {
    return true;
}

export function OpportunityRelationshipsPayloadFromJSON(json: any): OpportunityRelationshipsPayload {
    return OpportunityRelationshipsPayloadFromJSONTyped(json, false);
}

export function OpportunityRelationshipsPayloadFromJSONTyped(json: any, ignoreDiscriminator: boolean): OpportunityRelationshipsPayload {
    if (json == null) {
        return json;
    }
    return {
        
        'solutionIds': json['solutionIds'] == null ? undefined : json['solutionIds'],
        'outcomeIds': json['outcomeIds'] == null ? undefined : json['outcomeIds'],
        'feedbackIds': json['feedbackIds'] == null ? undefined : json['feedbackIds'],
    };
}

export function OpportunityRelationshipsPayloadToJSON(json: any): OpportunityRelationshipsPayload {
    return OpportunityRelationshipsPayloadToJSONTyped(json, false);
}

export function OpportunityRelationshipsPayloadToJSONTyped(value?: OpportunityRelationshipsPayload | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'solutionIds': value['solutionIds'],
        'outcomeIds': value['outcomeIds'],
        'feedbackIds': value['feedbackIds'],
    };
}